pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "washedmcp"
version = "0.1.0"
description = "Token-optimized semantic code search MCP server for AI coding assistants"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.14"
authors = [
    {name = "Pratham"}
]
keywords = ["mcp", "code-search", "semantic-search", "chromadb", "ai", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "chromadb~=0.4.0",
    "mcp~=1.0.0",
    "sentence-transformers~=2.2.0",
    "tree-sitter~=0.24.0",
    "tree-sitter-python~=0.23.0",
    "tree-sitter-javascript~=0.23.0",
    "tree-sitter-typescript~=0.23.0",
    "python-dotenv~=1.0.0",
]

[project.optional-dependencies]
summarize = ["anthropic~=0.18.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

[project.scripts]
washedmcp = "washedmcp.mcp_server:main"

[project.urls]
Homepage = "https://github.com/clarsbyte/washedmcp"
Repository = "https://github.com/clarsbyte/washedmcp"

[tool.setuptools.packages.find]
where = ["."]
include = ["washedmcp*"]

[tool.setuptools.package-data]
washedmcp = ["py.typed"]

# ============== Pytest Configuration ==============

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# ============== Coverage Configuration ==============

[tool.coverage.run]
source = ["washedmcp"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/mcp_server.py",
    "*/cli.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
precision = 2
fail_under = 60

[tool.coverage.html]
directory = "htmlcov"
